requirements.txt
absl-py==1.2.0+computecanada
anndata==0.8.0+computecanada
autopep8==1.7.0
cachetools==5.2.0+computecanada
certifi==2022.6.15+computecanada
charset-normalizer==2.1.1+computecanada
cycler==0.11.0+computecanada
easydict==1.9+computecanada
fonttools==4.34.4+computecanada
google-auth==2.11.0+computecanada
google-auth-oauthlib==0.4.6+computecanada
grpcio==1.47.0+computecanada
h5py==3.6.0+computecanada
idna==3.3+computecanada
importlib-metadata==4.12.0+computecanada
joblib==1.1.0+computecanada
kiwisolver==1.3.2+computecanada
llvmlite==0.38.1+computecanada
Markdown==3.4.1+computecanada
MarkupSafe==2.1.1+computecanada
matplotlib==3.5.2+computecanada
natsort==8.1.0+computecanada
networkx==2.8.5+computecanada
numba==0.55.2+computecanada
numexpr==2.8.1+computecanada
numpy==1.22.2+computecanada
oauthlib==3.2.0+computecanada
packaging==21.3+computecanada
pandas==1.4.1+computecanada
patsy==0.5.2+computecanada
Pillow==9.2.0+computecanada
protobuf==3.19.4+computecanada
pyasn1==0.4.8+computecanada
pyasn1-modules==0.2.8+computecanada
pycodestyle==2.9.1
pynndescent==0.5.2+computecanada
pyparsing==3.0.9+computecanada
python-dateutil==2.8.2+computecanada
pytz==2022.1+computecanada
PyYAML==6.0+computecanada
requests==2.28.1+computecanada
requests-oauthlib==1.3.1+computecanada
rsa==4.9+computecanada
scanpy==1.8.2+computecanada
scikit-learn==1.1.1+computecanada
scipy==1.8.0+computecanada
seaborn==0.11.2+computecanada
sinfo==0.3.1+computecanada
six==1.16.0+computecanada
statsmodels==0.13.2+computecanada
stdlib-list==0.7.0+computecanada
tables==3.7.0+computecanada
tabulate==0.8.10+computecanada
tensorboard==2.10.0+computecanada
tensorboard-data-server==0.6.1+computecanada
tensorboard-plugin-wit==1.8.1+computecanada
threadpoolctl==3.1.0+computecanada
toml==0.10.2+computecanada
torch==1.12.0+computecanada
tqdm==4.64.0+computecanada
typing-extensions==4.3.0+computecanada
umap-learn==0.5.1+computecanada
urllib3==1.26.12+computecanada
Werkzeug==2.2.2+computecanada
zipp==3.8.1+computecanada

